Fix an error with GCC 5.4

Follow-up to PR #18121 to resolve an error that occurs with GCC 5.4.

Trivial and not reviewed.

+++ compiler/next/include/chpl/types/QualifiedType.h
@@ -100,7 +100,7 @@ class QualifiedType {
     other.param_ = tmpParam;
   }
   size_t hash() const {
-    size_t h1 = chpl::hash(kind_);
+    size_t h1 = chpl::hash((unsigned) kind_);
     size_t h2 = chpl::hash(type_);
     size_t h3 = chpl::hash(param_);
